Do massive rearrangement of context tree stuff.
This involved decoupling the context tree itself from its presentation in the
NSOutlineView, and turning the ContextsDataSource class into a ContextTree
singleton class. This removed a few IB outlets as well, and simplified some
code.

This was only done to the English nib, though, so the nib changes will need to
be ported across to the other, non-English nibs.